Move ShimBundle into its own top-level class.
