+++ solstice/src/main/java/dev/equo/solstice/ShimStorage.java
@@ -34,7 +34,7 @@ class ShimStorage {
 		}
 	}
 
-	File getDataFileBundle(BundleContextSolstice.ShimBundle bundle, String filename) {
+	File getDataFileBundle(ShimBundle bundle, String filename) {
 		if (configDir == null) {
 			return null;
 		}
